feat - Add Product Create And View List Pages

+++ src/1.Domain/AYweb.Domain/Models/Product/Entities/Feature.cs
@@ -10,7 +10,7 @@ public class Feature : Entity<long>
 
     public string Value { get; private set; }
 
-    public int ProductId { get; private set; }
+    public long ProductId { get; private set; }
 
     public Product Product { get; private set; } 
     #endregion
